JumpScroll: new macro
Saves/loads subtitle grid scrollbar position.
Requires Aegisub with LuaJIT and win32/64 (r8238 or v3.2 and newer).
Number of "memory spots" for positions is set in the macro,
*jumpscroll_max = 3* by default.
Assign hotkeys to use it effectively e.g. Ctrl-F1...F3 to save,
Shift-F1...F3 to jump.
Currently it doesn't save these positions to a file so it's
session-only.